backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
langchain
langgraph
pytest
pytest-asyncio
python-dotenv
